switch to vitest and convert test to typescript
